requirements.txt
PyTMX==3.32
pathfinding==1.0.10
pygame-ce==2.5.0
